fix(search): fixing pages not loading after search
When search results lead to the `0.19` version, the page is no correctly
loaded. What happens is that the search result shows "page not found"
and only after refresh, the page is loaded.

Using the `externalUrlRegex to load all the searches related to the
`0.19` version in a separate window using `window.location.href` results
in the search on `0.19` to load correctly.

The drawback of this is that search results for `0.19` load a bit slower
because they need to render in a separate window.