Improved PRO label drawing, and fixed a display bug
The PRO label is now applied to the `::after` pseudo-selector of the
actual `.pro` element. This is to prevent a collision with Docusaurus'
own `::after`, which draws the submenus generally.

This affects the way that the PRO label is positioned (added comments
for future generations).

Additionally, I have implemented some changes for a more consistent look
for the PRO tags across different sections (though, some small and
mysterious quirks of style do remain)